requirements.txt
transformers
streamlit
pillow
plotly
chromadb
pypdf
pyautogen
tiktoken
sentence_transformers
"pyautogen[retrievechat]~=0.2.0b5"
"flaml[automl]"